Generic architecture for Quiz Controllers, quiz controllers migrated from storyboard to xib
