+++ Stepic/QuizViewController.xib
@@ -1,0 +1,88 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="9531" systemVersion="15E27e" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="9529"/>
+        <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="QuizViewController" customModule="Stepic" customModuleProvider="target">
+            <connections>
+                <outlet property="containerView" destination="HO0-Wd-0kw" id="BZi-zt-KtT"/>
+                <outlet property="sendButton" destination="yPX-01-eq5" id="V3v-9d-GUi"/>
+                <outlet property="statusImageView" destination="kBi-W4-vlY" id="TNt-3y-ooZ"/>
+                <outlet property="statusLabel" destination="hqL-ha-NG2" id="MJb-Y4-DeC"/>
+                <outlet property="statusViewHeight" destination="A6M-Cf-bMG" id="HUz-2F-ZAV"/>
+                <outlet property="view" destination="637-LV-Ael" id="kMB-Lb-wbK"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="637-LV-Ael">
+            <rect key="frame" x="0.0" y="0.0" width="600" height="355"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="I3R-Pc-iTV">
+                    <rect key="frame" x="-12" y="299" width="624" height="0.0"/>
+                    <subviews>
+                        <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="kBi-W4-vlY">
+                            <rect key="frame" x="16" y="-10" width="20" height="20"/>
+                            <constraints>
+                                <constraint firstAttribute="height" constant="20" id="6pa-6n-RUY"/>
+                                <constraint firstAttribute="width" constant="20" id="HQz-q8-2Oq"/>
+                            </constraints>
+                        </imageView>
+                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" verticalCompressionResistancePriority="749" ambiguous="YES" misplaced="YES" text="Status" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="hqL-ha-NG2">
+                            <rect key="frame" x="44" y="-6" width="572" height="0.0"/>
+                            <fontDescription key="fontDescription" type="system" pointSize="14"/>
+                            <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
+                            <nil key="highlightedColor"/>
+                        </label>
+                    </subviews>
+                    <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="calibratedWhite"/>
+                    <constraints>
+                        <constraint firstItem="hqL-ha-NG2" firstAttribute="top" secondItem="I3R-Pc-iTV" secondAttribute="top" priority="750" constant="6" id="0AG-fk-f5J"/>
+                        <constraint firstAttribute="height" id="A6M-Cf-bMG"/>
+                        <constraint firstAttribute="trailing" secondItem="hqL-ha-NG2" secondAttribute="trailing" constant="8" id="MtX-Ua-XZ7"/>
+                        <constraint firstItem="kBi-W4-vlY" firstAttribute="leading" secondItem="I3R-Pc-iTV" secondAttribute="leading" constant="16" id="UC9-Tf-tkc"/>
+                        <constraint firstItem="kBi-W4-vlY" firstAttribute="centerY" secondItem="I3R-Pc-iTV" secondAttribute="centerY" id="WXW-tl-3Ap"/>
+                        <constraint firstAttribute="bottom" secondItem="hqL-ha-NG2" secondAttribute="bottom" priority="750" constant="6" id="e1M-C9-bUs"/>
+                        <constraint firstItem="hqL-ha-NG2" firstAttribute="leading" secondItem="kBi-W4-vlY" secondAttribute="trailing" constant="8" id="zmf-Pf-8Il"/>
+                    </constraints>
+                </view>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="HO0-Wd-0kw" userLabel="container view">
+                    <rect key="frame" x="-12" y="0.0" width="624" height="291"/>
+                    <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="calibratedWhite"/>
+                </view>
+                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="yPX-01-eq5">
+                    <rect key="frame" x="8" y="307" width="584" height="40"/>
+                    <color key="backgroundColor" red="0.40000000600000002" green="0.80000001190000003" blue="0.40000000600000002" alpha="1" colorSpace="calibratedRGB"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="40" id="Pts-9U-YKn"/>
+                    </constraints>
+                    <state key="normal" title="Send">
+                        <color key="titleColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                    </state>
+                    <connections>
+                        <action selector="sendButtonPressed:" destination="-1" eventType="touchUpInside" id="VdX-f1-y2x"/>
+                    </connections>
+                </button>
+            </subviews>
+            <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+            <constraints>
+                <constraint firstItem="HO0-Wd-0kw" firstAttribute="leading" secondItem="637-LV-Ael" secondAttribute="leadingMargin" constant="-20" id="12n-C3-YYg"/>
+                <constraint firstItem="HO0-Wd-0kw" firstAttribute="top" secondItem="637-LV-Ael" secondAttribute="top" id="60n-bJ-YYh"/>
+                <constraint firstAttribute="trailingMargin" secondItem="I3R-Pc-iTV" secondAttribute="trailing" constant="-20" id="91e-vo-OtM"/>
+                <constraint firstItem="I3R-Pc-iTV" firstAttribute="leading" secondItem="637-LV-Ael" secondAttribute="leadingMargin" constant="-20" id="Gli-iz-mgq"/>
+                <constraint firstItem="I3R-Pc-iTV" firstAttribute="top" secondItem="HO0-Wd-0kw" secondAttribute="bottom" constant="8" id="SeL-c6-Aw2"/>
+                <constraint firstAttribute="trailingMargin" secondItem="yPX-01-eq5" secondAttribute="trailing" id="SyG-E5-BxU"/>
+                <constraint firstAttribute="leadingMargin" secondItem="yPX-01-eq5" secondAttribute="leading" id="YbC-Ag-TK4"/>
+                <constraint firstAttribute="trailingMargin" secondItem="HO0-Wd-0kw" secondAttribute="trailing" constant="-20" id="dgt-gf-5BW"/>
+                <constraint firstItem="yPX-01-eq5" firstAttribute="top" secondItem="I3R-Pc-iTV" secondAttribute="bottom" constant="8" id="gu1-58-epv"/>
+                <constraint firstAttribute="bottom" secondItem="yPX-01-eq5" secondAttribute="bottom" constant="8" id="xye-9J-7EX"/>
+            </constraints>
+            <nil key="simulatedStatusBarMetrics"/>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+            <point key="canvasLocation" x="282" y="-176.5"/>
+        </view>
+    </objects>
+</document>
