+++ Stepic/UnknownTypeQuizViewController.xib
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="9531" systemVersion="15E27e" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="9529"/>
+        <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="UnknownTypeQuizViewController" customModule="Stepic" customModuleProvider="target">
+            <connections>
+                <outlet property="solveOnTheWebsiteButton" destination="g9f-OI-IGK" id="Ndg-gv-JCY"/>
+                <outlet property="view" destination="Hwy-1O-t6N" id="f4C-Ty-MqG"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="Hwy-1O-t6N">
+            <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="g9f-OI-IGK">
+                    <rect key="frame" x="8" y="8" width="584" height="584"/>
+                    <color key="backgroundColor" red="0.40000000600000002" green="0.80000001190000003" blue="0.40000000600000002" alpha="1" colorSpace="calibratedRGB"/>
+                    <state key="normal" title="Solve on the website">
+                        <color key="titleColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                    </state>
+                    <connections>
+                        <action selector="solveOnTheWebsitePressed:" destination="-1" eventType="touchUpInside" id="BEv-sh-X0T"/>
+                    </connections>
+                </button>
+            </subviews>
+            <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+            <constraints>
+                <constraint firstItem="g9f-OI-IGK" firstAttribute="top" secondItem="Hwy-1O-t6N" secondAttribute="top" constant="8" id="0IH-8Q-skN"/>
+                <constraint firstAttribute="leadingMargin" secondItem="g9f-OI-IGK" secondAttribute="leading" id="Kzu-mM-l63"/>
+                <constraint firstAttribute="trailingMargin" secondItem="g9f-OI-IGK" secondAttribute="trailing" id="QUb-c1-xjq"/>
+                <constraint firstAttribute="bottom" secondItem="g9f-OI-IGK" secondAttribute="bottom" constant="8" id="lYZ-ss-FSo"/>
+            </constraints>
+            <nil key="simulatedStatusBarMetrics"/>
+            <nil key="simulatedTopBarMetrics"/>
+        </view>
+    </objects>
+</document>
